[log] Move tree related methods to logger_tree #33
- they are not related to core logic of logger
- they are mainly used to keep track of where all the loggers are and
allow user to have fine grained control of logger from different packages
